Move common text fixtures to generators test dir
